remove PledgeViewModel and Tests
